backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
python-dotenv==1.0.1
transformers==4.42.0
torch==2.3.0
keybert==0.8.5
scikit-learn==1.5.0
sentencepiece==0.2.0
spacy==3.7.5
httpx==0.27.0
